backend/requirements.txt
# Core Django
Django==5.2.4
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-filter==24.3

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
django-guardian==2.4.0
cryptography==42.0.8

# Database
# psycopg2-binary==2.9.9  # Commented out due to Python 3.13 compatibility issues
redis==5.0.8
django-redis==5.4.0

# File handling & Media
Pillow==10.4.0
django-storages==1.14.4
boto3==1.34.162

# Validation & Utilities
django-phonenumber-field==8.0.0
phonenumbers==8.13.45
python-decouple==3.8
celery==5.3.6

# API Documentation
drf-spectacular==0.27.2

# Development & Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.1
django-debug-toolbar==4.4.6

# Production
gunicorn==23.0.0
whitenoise==6.7.0

# Monitoring & Logging
sentry-sdk==2.13.0
django-extensions==3.2.3 